ISIS-2717: converts ExecutionParametersService into an interface rather than an impl ...
... with ExecutionParamterServiceAutoConfiguration providing the fallback.
