--- testing/fixtures/adoc/modules/fixtures/pages/services/ExecutionParametersService.adoc
@@ -1,27 +1,0 @@
-= `ExecutionParametersService`
-
-:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-:page-partial:
-
-CAUTION: TODO: this needs to be converted into a hook for refguide-index
-
-
-The `ExecutionParametersService` is used by the framework simply to instantiate the `ExecutionParameters` object.
-The `ExecutionParameters` object in turn is responsible for parsing the string parameter passed when executing fixtures through the UI to the xref:testing:fixtures:services/FixtureScripts.adoc[FixtureScripts] domain service.
-
-== API & Implementation
-
-The API and implementation of this service is simply:
-
-[source,java]
-----
-public class ExecutionParametersService {
-    public ExecutionParameters newExecutionParameters(final String parameters) {
-        return new ExecutionParameters(parameters);
-    }
-}
-----
-
-
-
-
