+++ testing/fixtures/applib/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  org.apache.isis.testing.fixtures.applib.fixturespec.FixtureScriptsSpecificationProviderAutoConfiguration
+  org.apache.isis.testing.fixtures.applib.fixturespec.FixtureScriptsSpecificationProviderAutoConfiguration,\
+  org.apache.isis.testing.fixtures.applib.fixturescripts.ExecutionParametersServiceAutoConfiguration
